eye tuner detector web 210708.v2 commit

+++ flask/project/static/js/common.js
@@ -1,3 +1,28 @@
+/** Commonly-used functions */
+
+function hypot(...val) {
+  return Math.sqrt(val.reduce((previousValue, currentValue) => (previousValue + Math.pow(currentValue, 2)), 0));
+}
+
+function distance(p1, p2) {
+  return hypot(p2.x - p1.x, p2.y - p1.y);
+}
+
+function ensureLoadedFaceApi(uri, callback=null) {
+  if (!!callback) {
+    Promise.all([
+      faceapi.nets.tinyFaceDetector.loadFromUri(uri),
+      faceapi.nets.faceLandmark68Net.loadFromUri(uri),
+      faceapi.nets.faceRecognitionNet.loadFromUri(uri),
+      faceapi.nets.faceExpressionNet.loadFromUri(uri)
+    ]).then(callback)
+  } else {
+    faceapi.nets.tinyFaceDetector.loadFromUri(uri)
+    faceapi.nets.faceLandmark68Net.loadFromUri(uri)
+    faceapi.nets.faceRecognitionNet.loadFromUri(uri)
+    faceapi.nets.faceExpressionNet.loadFromUri(uri)
+  }
+}
 
 function getEyeLengthByCoordinates(eyeCoordinates) {
   let left_point, right_point, top_point, bot_point
@@ -23,57 +48,59 @@ function getMaxMinCoordinates(eyeCoordinates) {
 
 function getEyeMask(width, height, eyeCoordinates) {
 
-  let cvs = document.createElement("CANVAS");
-  cvs.style.display = "none";
-  cvs.width = width;
-  cvs.height = height;
+  let cvs;
+  let maskRGBA;
 
-  let ctx = cvs.getContext('2d');
-  ctx.fillStyle = "black";
-  ctx.fillRect(0, 0, width, height)
+  try {
+    cvs = document.createElement('CANVAS');
+    cvs.style.display = 'none';
+    cvs.width = width;
+    cvs.height = height;
 
-  let [f, ...r] = eyeCoordinates
-  ctx.beginPath();
-  ctx.moveTo(f.x, f.y);
-  r.forEach(p => ctx.lineTo(p.x, p.y));
-  ctx.closePath();
-  ctx.fillStyle = "white";  // ctx.strokeStyle
-  ctx.fill();  // ctx.fill
+    let ctx = cvs.getContext('2d');
+    ctx.fillStyle = 'black';
+    ctx.fillRect(0, 0, width, height);
 
-  let maskRGBA = cv.imread(cvs);
-  let mask = new cv.Mat();
-  cv.cvtColor(maskRGBA, mask, cv.COLOR_RGBA2GRAY, 0);
+    let [f, ...r] = eyeCoordinates;
+    ctx.beginPath();
+    ctx.moveTo(f.x, f.y);
+    r.forEach(p => ctx.lineTo(p.x, p.y));
+    ctx.closePath();
+    ctx.fillStyle = 'white';  // ctx.strokeStyle
+    ctx.fill();  // ctx.fill
 
-  cvs.remove()
-  maskRGBA.delete()
+    maskRGBA = cv.imread(cvs);
+    let mask = new cv.Mat();
+    cv.cvtColor(maskRGBA, mask, cv.COLOR_RGBA2GRAY, 0);
 
-  return mask;
+    return mask;
+
+  } finally {
+    if (cvs) cvs.remove();
+    if (maskRGBA) maskRGBA.delete();
+  }
 
 }
 
 function getGrayEyeFromCanvas(canvas, eyeCoordinates) {
 
-  let gray = cv.imread(canvas);
-  cv.cvtColor(gray, gray, cv.COLOR_RGBA2GRAY, 0);
+  let gray, mask;
 
-  let mask = getEyeMask(canvas.width, canvas.height, eyeCoordinates);
-  cv.bitwise_and(gray, gray, gray, mask)
+  try {
 
-  let [min_x, min_y, max_x, max_y] = getMaxMinCoordinates(eyeCoordinates).map(Math.round)
-  let rect = new cv.Rect(min_x, min_y, max_x - min_x, max_y - min_y)
-  let eye = gray.roi(rect)
+    gray = cv.imread(canvas);
+    cv.cvtColor(gray, gray, cv.COLOR_RGBA2GRAY, 0);
 
-  gray.delete();
-  mask.delete();
-  return eye;
+    mask = getEyeMask(canvas.width, canvas.height, eyeCoordinates);
+    cv.bitwise_and(gray, gray, gray, mask);
+
+    let [min_x, min_y, max_x, max_y] = getMaxMinCoordinates(eyeCoordinates).map(Math.round);
+    let rect = new cv.Rect(min_x, min_y, max_x - min_x, max_y - min_y);
+    return gray.roi(rect);
+
+  } finally {
+    if (gray) gray.delete();
+    if (mask) mask.delete();
+  }
 
 }
-
-function hypot(...val) {
-  return Math.sqrt(val.reduce((previousValue, currentValue) => (previousValue + Math.pow(currentValue, 2)), 0));
-}
-
-function distance(p1, p2) {
-  return hypot(p2.x - p1.x, p2.y - p1.y);
-}
-
